util_adcfifo: Fix the write logic of the FIFO
The write address turned around at the end of the FIFO, losing the last
several beats of the FIFO. This way, the DMA could not read out all the
saved data. Change the logic to make sure that the write address counter
does not turns around. Also make sure that the address, data and write
enable signals are correctly aligned.